Changed definition of arrays
In the new version, the arrays contain only distinct (unique) values. For example, if in a same transaction the token ID is transfered twice, nft_token_ids_array will only have one value (but the number_of_items will be 2). Parts of the documentation was modified to reflect theses changes